backend/python-apis/requirements.txt
numpy==1.24.3
scipy==1.10.1
pandas==2.0.3
reportlab==4.0.4
openpyxl==3.1.2
apscheduler==3.10.4
scikit-learn==1.3.0
psycopg2-binary==2.9.7
aiohttp==3.8.5
asyncio==3.4.3
requests==2.31.0
